Add example with nesting and resource groups
This is loosely based on the MIT-licensed WordPress example Heat
tutorial from Miguel Grinberg, available at
https://github.com/miguelgrinberg/heat-tutorial/.

It deploys a scaleable WordPress environment with an NFS server, a
MySQL server, and a configurable number of WordPress front ends, using
Cinder volumes as the persistent data store.

It uses some crudely idempotent shell scripts invoked from cloud-init
to automate the deployment of MySQL, NFS, and WordPress. Needless to
say, most real-world environments would likely be better served by
Ansible -- however, the point of this template is to show what can
be done using as few external tools as possible.

It makes use of Neutron LBaaS to load-balance the Apache servers
running WordPress, and to enable SSH access to the backend MySQL host
(which can then be used as a jump host to access all other Nova
instances).

Any instance in this stack can be safely deleted, including the NFS
and MySQL instances. As long as the volumes continue to live, a simple
stack check, followed by a stack update with unchanged parameters,
will bring WordPress right back.